+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 ### Added
 
 - First Release
+- Added context menu actions to insert an asset as a paint layer with `Document.createNode`.
+- Added context menu action to insert an asset as a file layer with `Document.createFileLayer`.
+- Added context menu action to duplicate an asset file with a prompted destination name.